@@ -1,23 +1,29 @@
 {
   "name": "ezbilling-ui",
-  "version": "0.1.0",
+  "version": "1.0.0",
   "private": true,
   "dependencies": {
-    "@reduxjs/toolkit": "^1.9.1",
-    "@testing-library/jest-dom": "^5.11.10",
-    "@testing-library/react": "^11.2.6",
-    "@testing-library/user-event": "^12.8.3",
-    "axios": "^1.5.0",
-    "bootstrap": "4.6.0",
-    "framer-motion": "^4.1.3",
-    "react": "^17.0.2",
+    "@reduxjs/toolkit": "^1.9.5",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.0.1",
+    "@testing-library/user-event": "^14.1.1",
+    "axios": "^0.27.2",
+    "bootstrap": "^5.3.1",
+    "fetch-retry": "^1.1.1",
+    "formik": "^2.4.3",
+    "react": "^18.2.0",
     "react-bootstrap": "^2.8.0",
-    "react-icons": "^4.2.0",
-    "react-redux": "^7.2.3",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.1.2",
+    "react-router-dom": "^6.4.0",
     "react-scripts": "5.0.1",
-    "react-toastify": "8.0.0",
-    "redux": "^4.0.5",
-    "web-vitals": "^1.1.1"
+    "react-toastify": "^9.1.3",
+    "react-validation": "^3.0.7",
+    "redux": "^4.2.1",
+    "redux-devtools-extension": "^2.13.9",
+    "redux-thunk": "^2.4.2",
+    "validator": "^13.7.0",
+    "yup": "^1.2.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -42,8 +48,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "engines": {
-    "node": ">=14.0.0"
   }
 }
