@@ -1,38 +1,22 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
+import React from "react";
+import { createRoot } from "react-dom/client";
+import { BrowserRouter } from "react-router-dom";
+import {store  }     from "./components/redux/store";
+import { Provider } from "react-redux";
 
-ReactDOM.render(
-  <React.StrictMode>
-    {/* <App /> */}
-  </React.StrictMode>,
-  document.getElementById('root')
+import App from "./App";
+import * as serviceWorker from "./serviceWorker";
+
+const container = document.getElementById("root");
+const root = createRoot(container);
+
+root.render(
+  <BrowserRouter>
+  <Provider store ={store}>
+    <App />
+  </Provider >
+  </BrowserRouter>
+
 );
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-reportWebVitals();
-
-// import React from 'react';
-// import ReactDOM from 'react-dom/client';
-// // import './index.css';
-// import App from './App';
-// import reportWebVitals from './reportWebVitals';
-// import configureStore from './components/redux/configureStore';
-// // import { Provider } from 'react-redux';
-// // 
-// const root = ReactDOM.createRoot(document.getElementById('root'));
-// root.render(
-//   <React.StrictMode>
-//     <Provider store={configureStore}>
-//     <App />
-//     </Provider>
-//   </React.StrictMode>
-// );
-
-// // If you want to start measuring performance in your app, pass a function
-// // to log results (for example: reportWebVitals(console.log))
-// // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-// reportWebVitals();
+serviceWorker.unregister();
