@@ -3,35 +3,38 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 import Login from "./components/login/Login";
 import UserRoutes from "./components/userRoutes/userRoutes";
 import Dashboard from "./components/dashboard/dashboard";
-// import {BrowserRouter,
-//   Routes, // instead of "Switch"
-//   Route,
-// } from "react-router-dom";
+import {
+  BrowserRouter,
+  Routes, // instead of "Switch"
+  Route,
+} from "react-router-dom";
 import Navbar from "./components/navbar/EzNavbar";
-import  Company from "./components/Company";
-import'./App.css';
+import './App.css';
 import About from './components/about/About';
-import { bindActionCreators } from '@reduxjs/toolkit';
-import { actions } from './components/redux/AuthActions';
-const App=(props)=> {
+import { useDispatch, useSelector } from 'react-redux';
+import { doLogin } from './components/redux/loginSlice';
+import { Navigate, useNavigate } from "react-router-dom";
+const App = (props) => {
+
+  const dispatch = useDispatch();
+  const { isLoggedIn } = useSelector((state) => state.ezLogin);
 
   return (
-    <div>app</div>
-  //   <BrowserRouter>
-  //   <Navbar/>
-  //   <div className='routes'>
-  //   <Routes >
-  //     <Route path="/login" element={<Login {...props} />}/>
-  //     <Route path="/about" element={<About />}/>
-  //     <Route path ="/user" element={<UserRoutes/>}>
-  //       <Route path ="dashboard" element={<Dashboard/>}/>
-  //     </Route>
-  //   </Routes>
-  //   </div>
-  // </BrowserRouter> 
+
+    <div>
+      <Navbar />
+      <div className='routes'>
+        <Routes >
+          <Route path="/login" element={<Login {...props} />} />
+          <Route path="/about" element={<About />} />
+          <Route path="/user" element={<UserRoutes />}>
+            <Route path="dashboard" element={<Dashboard />} />
+          </Route>
+        </Routes>
+      </div>
+    </div>
   );
 }
 
-
-
 export default App;
+      
